written workaround that fixes problem during kotti startup (populator phase). No request available at this time. To be fixed better definitively

+++ kotti_es/events.py
@@ -18,31 +18,34 @@ DELETE_CODE = -1
 
 def _after_insert(mapper, connection, target):
     request = get_current_request()
-    if IContent.providedBy(target):
-        request._index_list = [(target, INSERT_CODE)]
+    if request:
+        if IContent.providedBy(target):
+            request._index_list = [(target, INSERT_CODE)]
 
 
 def _after_delete(mapper, connection, target):
     request = get_current_request()
-    if IContent.providedBy(target):
-        if not hasattr(request, '_index_list'):
-            request._index_list = [(target, DELETE_CODE)]
-        else:
-            request._index_list.append((target, DELETE_CODE))
+    if request:
+        if IContent.providedBy(target):
+            if not hasattr(request, '_index_list'):
+                request._index_list = [(target, DELETE_CODE)]
+            else:
+                request._index_list.append((target, DELETE_CODE))
 
 
 def _after_commit(session):
     request = get_current_request()
-    index_list = getattr(request, '_index_list', [])
-    es_client = get_client(request)
-    for target, operation in index_list:
-        if operation == INSERT_CODE:
-            es_client.index_object(target, immediate=True)
-        elif operation == DELETE_CODE:
-            try:
-                es_client.delete_object(target, immediate=True)
-            except NotFoundError:
-                pass
+    if request:
+        index_list = getattr(request, '_index_list', [])
+        es_client = get_client(request)
+        for target, operation in index_list:
+            if operation == INSERT_CODE:
+                es_client.index_object(target, immediate=True)
+            elif operation == DELETE_CODE:
+                try:
+                    es_client.delete_object(target, immediate=True)
+                except NotFoundError:
+                    pass
 
 
 def wire_sqlalchemy():  # pragma: no cover
